Updated the capabilities of the bash script

+++ CreateForgeServer.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
 # Default values
-AUTO_APPROVE=''
+SKIP_MODPACK_DOWNLOAD=''
+SKIP_SERVER_INITIALIZATION=''
+URL_MODPACK=''
 OUTPUT_DIR='OUTPUT'
+AUTO_APPROVE=''
 RAM_MIN='1024'
 RAM_MAX='4096'
 LOG='true'
@@ -15,6 +18,9 @@ showHelp()
     echo "  ${0} [-h / --help] [-o / --out-dir] NEW_SERVER"
     echo '  [-n / --ram-min] 1024 [-x / --ram-max] 4096'
     echo '  [-q / --quiet] [-a / --approve]'
+    echo '  [-fv / --forge-version] [-mv / --minecraft-version]'
+    echo '  [-mp / --modpack] [-sm / --skip-modpack-downoad]'
+    echo '  [-si / --skip-initialization]'
     echo ''
     echo 'Options:'
     echo '  -h, --help'
@@ -27,6 +33,17 @@ showHelp()
     echo '    Maximum amount of ram to be used with java'
     echo '  -a, --approve'
     echo '    Auto approve eula'
+    echo '  -fv, --forge-version'
+    echo '    The version of Forge to download'
+    echo '  -mv, --minecraft-version'
+    echo '    The version of Minecraft to download'
+    echo '  -m, --modpack'
+    echo '    The URL to fetch the modpack at'
+    echo '  -smd, --skip-modpack-download'
+    echo '    Skip the download of the modpack, this is useful'
+    echo '    for if the modpack is already local'
+    echo '  -si, --skip-initialization'
+    echo '    Skip the initialization of the modpack with forge'
     echo ''
 }
 
@@ -56,6 +73,26 @@ do
     -a | --approve )
         AUTO_APPROVE='true'
     ;;
+    -fv | --forge-version )
+        shift
+        FORGE_VERSION=${1}
+    ;;
+    -mv | --minecraft-version )
+        shift
+        MC_VERSION=${1}
+    ;;
+    -m | --modpack )
+        shift
+        URL_MODPACK=${1}
+    ;;
+    -smd | --skip-modpack-download )
+        shift
+        SKIP_MODPACK_DOWNLOAD='true'
+    ;;
+    -si | --skip-initialization )
+        shift
+        SKIP_SERVER_INITIALIZATION='true'
+    ;;
     * )
         echo Unknown option \"${1}\"
         exit 1
@@ -63,10 +100,23 @@ do
     shift
 done
 
+# Validations
+if [ !${MC_VERSION} ]
+then
+    echo 'Missing Minecraft version argument, stopping process'
+    exit 1
+elif [ !${FORGE_VERSION} ]
+then
+    echo 'Missing Forge version argument, stopping process'
+    exit 1
+elif [ ${URL_MODPACK} && ${SKIP_MODPACK_DOWNLOAD} ]
+then
+    echo 'Either a modpack URL has to be specified, or skipped with (-smd/--skip-modpack-download), stopping process'
+    exit 1
+fi
+
 # Setup
-URL_FORGE_INSTALLER='https://files.minecraftforge.net/maven/net/minecraftforge/forge/1.12.2-14.23.5.2847/forge-1.12.2-14.23.5.2847-installer.jar'
-URL_MODPACK='https://media.forgecdn.net/files/2836/138/RLCraft+Server+Pack+1.12.2+-+Beta+v2.8.1.zip'
-OLD_IFS=${IFS}
+URL_FORGE_INSTALLER="https://maven.minecraftforge.net/net/minecraftforge/forge/${MC_VERSION}-${FORGE_VERSION}/forge-${MC_VERSION}-${FORGE_VERSION}-installer.jar"
 DIR="$(pwd)/${OUTPUT_DIR}"
 if [ -d ${OUTPUT_DIR} ]
 then
@@ -77,16 +127,11 @@ mkdir ${OUTPUT_DIR}
 cd ${OUTPUT_DIR}
 
 # Get version numbers
-IFS='-'
-read -ra ARR <<< ${URL_FORGE_INSTALLER}
-FORGE_MCVERSION="${ARR[-3]}"
-FORGE_VERSION="${ARR[-2]}"
-IFS=${OLD_IFS}
-FORGE_INSTALLER_FILE="${DIR}/forge-${FORGE_MCVERSION}-${FORGE_VERSION}-installer.jar"
-FORGE_UNIVERSAL_FILE="${DIR}/forge-${FORGE_MCVERSION}-${FORGE_VERSION}-universal.jar"
+FORGE_INSTALLER_FILE="${DIR}/forge-${MC_VERSION}-${FORGE_VERSION}-installer.jar"
+FORGE_UNIVERSAL_FILE="${DIR}/forge-${MC_VERSION}-${FORGE_VERSION}-universal.jar"
 
 # Get Forge installer
-echo '> Fetching Forge installer'
+echo "> Fetching Forge installer - ${URL_FORGE_INSTALLER}"
 wget -q ${URL_FORGE_INSTALLER}
 echo '> Unzipping Forge installer'
 unzip -q ${FORGE_INSTALLER_FILE}
@@ -102,21 +147,27 @@ rm ${FORGE_INSTALLER_FILE} ${FORGE_INSTALLER_FILE}.log forge-${FORGE_MCVERSION}-
 echo '> Deleted installer and log'
 
 # Pull and unzip modpack
-MODPACK_ZIP='modpack.zip'
-echo '> Fetching modpack'
-wget -q -O ${MODPACK_ZIP} ${URL_MODPACK} # look at --progress=TYPE for progress
-echo '> Unzipping modpack'
-unzip -q -d tmp ${MODPACK_ZIP}
-cp -a tmp/. ./
-rm -r tmp ${MODPACK_ZIP}
+if [ !${SKIP_MODPACK_DOWNLOAD} ]
+then
+    MODPACK_ZIP='modpack.zip'
+    echo '> Fetching modpack'
+    wget -q -O ${MODPACK_ZIP} ${URL_MODPACK} # look at --progress=TYPE for progress
+    echo '> Unzipping modpack'
+    unzip -q -d tmp ${MODPACK_ZIP}
+    cp -a tmp/. ./
+    rm -r tmp ${MODPACK_ZIP}
+fi
 
 # Initialize server
-echo '> Initializing the server'
-if [ ${LOG} ]
+if [ !${SKIP_SERVER_INITIALIZATION} ]
 then
-    java -Xms${RAM_MIN}M -Xmx${RAM_MAX}M -jar ${FORGE_UNIVERSAL_FILE} nogui >> ../CreateForgeLog.txt
-else
-    java -Xms${RAM_MIN}M -Xmx${RAM_MAX}M -jar ${FORGE_UNIVERSAL_FILE} nogui > /dev/null 2>&1
+    echo '> Initializing the server'
+    if [ ${LOG} ]
+    then
+        java -Xms${RAM_MIN}M -Xmx${RAM_MAX}M -jar ${FORGE_UNIVERSAL_FILE} nogui >> ../CreateForgeLog.txt
+    else
+        java -Xms${RAM_MIN}M -Xmx${RAM_MAX}M -jar ${FORGE_UNIVERSAL_FILE} nogui > /dev/null 2>&1
+    fi
 fi
 
 # Handle eula
